qcacld-3.0: Update the AKM data type
Currently the AKM information is sent as uint16_t variable
this leads to truncation of 1byte in connecting event of
advanced connectivity logging.

Update the AKM data type from uint16_t to uint32_t.

CRs-Fixed: 3307515

+++ components/cmn_services/logging/inc/wlan_connectivity_logging.h
@@ -466,8 +466,8 @@ struct wlan_diag_connect {
 	char ssid[WLAN_SSID_MAX_LEN];
 	uint8_t bssid_hint[6];
 	uint16_t reason;
-	uint16_t akm;
-	uint16_t subtype;
+	uint32_t akm;
+	uint32_t subtype;
 	uint32_t freq;
 	uint32_t freq_hint;
 	uint32_t pairwise_cipher;
